refactor: extract AudioRecorderEngine.swift from AudioRecorderTypes.swift
Splits the `AudioRecorderEngine` @MainActor protocol + its
`AVAudioRecorder` conformance extension out of AudioRecorderTypes.swift.

Byte-identical move of lines 9-21. Both files import
@preconcurrency AVFoundation + Foundation.

AudioRecorderTypes.swift: 57 → 43 lines. Tests: 667.
