+++ Sources/BugNarrator/Services/AudioRecorderTypes.swift
@@ -6,20 +6,6 @@ struct RecordedAudio {
     let duration: TimeInterval
 }
 
-@MainActor
-protocol AudioRecorderEngine: AnyObject {
-    var delegate: (any AVAudioRecorderDelegate)? { get set }
-    var currentTime: TimeInterval { get }
-
-    func prepareToRecord() -> Bool
-    func record() -> Bool
-    func stop()
-}
-
-extension AVAudioRecorder: AudioRecorderEngine {}
-
-typealias AudioRecorderEngineFactory = (URL, [String: Any]) throws -> any AudioRecorderEngine
-
 enum AudioRecorderCaptureFormat: Equatable {
     case aacM4A
     case wavPCM
